Make named imports in type=module packages work.
In this article [1] I learned how node "interprets" and "understands"
commonjs files, which this file is (lib/hamjest.js). But in order for
named imports like `import {assertThat} from 'hamjest'` work, the export
must be simpler (simply said). That is what this commit does, it allows
to statically analyse the hamjest package and make it work when imported
as ESM.

[1]
https://simonplend.com/node-js-now-supports-named-imports-from-commonjs-modules-but-what-does-that-mean/